Move Cache to AOP
