+++ vertx-gaia/vertx-up/src/main/java/io/vertx/up/uca/jooq/JqWriter.java
@@ -28,7 +28,6 @@ import java.util.function.Function;
 class JqWriter {
 
     private transient final VertxDAO vertxDAO;
-    private transient final JqL1 cacheL1;
     private transient JqReader reader;
     private transient JqAnalyzer analyzer;
 
@@ -36,7 +35,6 @@ class JqWriter {
         this.vertxDAO = vertxDAO;
         this.analyzer = analyzer;
         this.reader = JqReader.create(vertxDAO, analyzer);
-        this.cacheL1 = JqL1.create(vertxDAO, analyzer);
     }
 
     static JqWriter create(final VertxDAO vertxDAO, final JqAnalyzer analyzer) {
@@ -75,10 +73,8 @@ class JqWriter {
 
     /* Sync insert operation: INSERT */
     // Cached
-    @L1
     <T> T insert(final T entity) {
         this.vertxDAO.insert(uuid(entity));
-        this.cacheL1.insertL1(entity);
         return entity;
     }
 
@@ -102,10 +98,8 @@ class JqWriter {
     }
 
     /* Sync insert operation: UPDATE */
-    @L1
     <T> T update(final T entity) {
         this.vertxDAO.update(entity);
-        this.cacheL1.updateL1(entity);
         return entity;
     }
 
@@ -141,7 +135,6 @@ class JqWriter {
     /* Sync delete operation: DELETE */
     <T> T delete(final T entity) {
         this.vertxDAO.delete(entity);
-        this.cacheL1.deleteL1(entity);
         return entity;
     }
