+++ vertx-gaia/vertx-up/src/main/java/io/vertx/up/uca/jooq/aop/L1.java
@@ -1,4 +1,4 @@
-package io.vertx.up.uca.jooq;
+package io.vertx.up.uca.jooq.aop;
 
 import java.lang.annotation.*;
