+++ vertx-gaia/vertx-up/zero-c-debug.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+mvn clean package install -DskipTests=true -Dmaven.javadoc.skip=true -Dmaven.compile.fork=true -T 1C
